@@ -27,12 +27,12 @@ const zerr = require('../util/zerr.js');
 const zfile = require('../util/file.js');
 const lpm_config = require('../util/lpm_config.js');
 const qw = require('../util/string.js').qw;
-const child_process = require('child_process');
 const uuid_v4 = require('uuid/v4');
 const sessions = require('./session.js');
 const Context = require('./context.js');
 const Router = require('./router.js');
 const Rules = require('./rules.js');
+const ReqProcess = require('./req_process.js');
 const assign = Object.assign, {SEC} = date.ms;
 const E = module.exports = Luminati;
 E.user_agent = 'luminati-proxy-manager/'+version;
@@ -162,6 +162,8 @@ function Luminati(opt, mgr){
     this.failure = {};
     this.requests_queue = [];
     this.throttle_queue = [];
+    this.req_process = new ReqProcess({listen_port: opt.listen_port,
+        log: opt.log});
     this.http_server = http.createServer((req, res, head)=>{
         if (this.is_overload())
             return this.send_overload(res, 'http_req');
@@ -177,6 +179,7 @@ function Luminati(opt, mgr){
     http_shutdown(this.http_server);
     this.socks_server = new Socks({port: opt.port||E.default.port,
         log: opt.log});
+    this.socks_server.on('error', err=>this.emit('error', err));
     if (opt.ssl)
     {
         this.authorization = {};
@@ -708,7 +711,7 @@ function*init_proxy_req(_this, req, res){
     };
     ctx.connect_headers = {
         'proxy-authorization': 'Basic '+
-            new Buffer(ctx.cred.username+':'+ctx.cred.password)
+            Buffer.from(ctx.cred.username+':'+ctx.cred.password)
             .toString('base64'),
     };
     if (!ctx.ext_proxy)
@@ -755,17 +758,7 @@ E.prototype._request = etask._fn(function*_request(_this, req, res, head){
                 'HTTPS connection to IP will be done from super agent');
         }
         const remote = req.socket.remoteAddress+':'+req.socket.remotePort;
-        const netstat_cmd = `netstat -antp 2>/dev/null | grep "${remote}"`;
-        let pid;
-        if (!lpm_config.is_win && req.socket.remoteAddress=='127.0.0.1')
-        {
-            try {
-                const stdout = child_process.execSync(netstat_cmd);
-                pid = stdout.toString().split('\n').filter(Boolean)
-                .map(s=>s.replace(/\s+/g, ' ').split(' '))
-                .find(r=>r[3]==remote);
-            } catch(e){ pid = ''; }
-        }
+        _this.req_process.add(req);
         _this._handle_usage_start(req);
         let resp;
         if (ctx.rules)
@@ -787,7 +780,8 @@ E.prototype._request = etask._fn(function*_request(_this, req, res, head){
             throw new Error('invalid_response');
         if (ctx.wait_bw)
             yield this.wait_ext(ctx.wait_bw);
-        resp.remote_address = pid ? `${remote}, ${pid[6]}` : remote;
+        resp.remote_address = req.process ?
+            `${remote}, ${req.process}` : remote;
         resp.uuid = req.ctx.uuid;
         const auth = username.parse(ctx.h_proxy_authorization);
         if (auth&&auth.tool=='proxy_tester')
