@@ -251,7 +251,7 @@ E.prototype.init = function(argv, run_config){
     this._defaults = conf._defaults;
     this.proxies = conf.proxies;
     this.mode = argv.mode;
-    this.loki = new Loki(argv.loki);
+    this.loki = new Loki(argv.loki, argv.log);
     if (!['root', 'normal', 'guest'].includes(this.mode))
     {
         console.log(`Unrecognized UI mode (${this.mode}); treating as guest`);
@@ -511,6 +511,7 @@ function*mgr_stop(_this, reason, force, restart){
     _this.is_running = false;
     yield _this.perr(restart ? 'restart' : 'exit', {reason});
     yield _this.stop_dbs();
+    yield _this.loki.save();
     if (reason!='config change')
         yield _this.save_config();
     if (reason instanceof Error)
@@ -1783,7 +1784,7 @@ E.prototype.test_api = function(req, res){
     });
     const password = this.proxies_running[port].config.password;
     const user = 'tool-proxy_tester';
-    const basic = new Buffer(user+':'+password).toString('base64');
+    const basic = Buffer.from(user+':'+password).toString('base64');
     opt.headers = opt.headers||{};
     opt.headers['proxy-authorization'] = 'Basic '+basic;
     if (+port)
@@ -1792,8 +1793,6 @@ E.prototype.test_api = function(req, res){
         if (this.proxies_running[port].opt.ssl)
             opt.ca = ssl.ca.cert;
         opt.headers = opt.headers||{};
-        opt.headers['x-hola-context'] = opt.headers['x-hola-context']
-            ||'PROXY TESTER TOOL';
     }
     request(opt, err=>{
         if (!err)
@@ -1896,8 +1895,6 @@ E.prototype.logs_resend = etask._fn(function*mgr_logs_resend(_this, req, res){
         };
         if (proxy.opt.ssl)
             opt.ca = ssl.ca.cert;
-        opt.headers['x-hola-context'] = opt.headers['x-hola-context']
-            ||'HAR VIEWER';
         request(opt);
     }
     res.send('ok');
@@ -2048,6 +2045,7 @@ E.prototype.get_settings = function(){
         api: this.argv.api,
         logs: this._defaults.logs,
         request_stats: this._defaults.request_stats,
+        dropin: this._defaults.dropin,
     };
 };
 
@@ -2271,7 +2269,7 @@ function*mgr_refresh_vips(_this, req, res){
         res.json({error: response.body});
 });
 
-E.prototype.shutdown = function(req, res){
+E.prototype.shutdown_api = function(req, res){
     res.json({result: 'ok'});
     this.stop(true);
 };
@@ -2355,7 +2353,37 @@ E.prototype.stats_clear_api = function(req, res){
     res.end();
 };
 
-E.prototype.sync_recent_stats = etask._fn(function*mgr_sync(_this){
+E.prototype.api_request = etask._fn(function*mgr_api_request(_this, opt){
+    const jar = _this.luminati_jar.jar;
+    yield etask.nfn_apply(request, [{url: _this.argv.api, jar}]);
+    const xsrf = (jar.getCookies(_this.argv.api).find(e=>
+        e.key=='XSRF-TOKEN')||{}).value;
+    return yield etask.nfn_apply(request, [{
+        method: opt.post||'GET',
+        url: opt.url,
+        qs: assign(_.pick(_this._defaults, qw`customer token`)),
+        jar,
+        json: true,
+        headers: {'X-XSRF-Token': xsrf},
+        form: opt.form,
+    }]);
+});
+
+E.prototype.sync_config_file = etask._fn(function*mgr_sync_config(_this){
+    if (!_this.logged_in)
+        return;
+    const res = yield _this.api_request({
+        url: `${_this.argv.api}/update_lpm_config`,
+        method: 'POST', form: {def: 1},
+    });
+    if (res.statusCode!=200)
+    {
+        _this._log.warn('Config file was not synced with mongodb, %s',
+            res.statusCode);
+    }
+});
+
+E.prototype.sync_recent_stats = etask._fn(function*mgr_sync_stats(_this){
     if (!_this.logged_in)
         return;
     const stats = _this.loki.stats_get();
@@ -2480,7 +2508,7 @@ E.prototype.create_api_interface = function(){
     app.get('/allocated_vips', this._api(this.allocated_vips_get, ['root']));
     app.post('/refresh_ips', this._api(this.refresh_ips_api, ['root']));
     app.post('/refresh_vips', this._api(this.refresh_vips_api, ['root']));
-    app.post('/shutdown', this._api(this.shutdown, ['root']));
+    app.post('/shutdown', this._api(this.shutdown_api, ['root']));
     app.post('/logout', this._api(this.logout_api));
     app.post('/upgrade', this._api(this.upgrade, ['root']));
     app.post('/restart', this._api(this.restart, ['root']));
