@@ -105,7 +105,7 @@ class Tracer {
         const user = Object.keys(opt).map(k=>`${k}-${opt[k]}`).join('-');
         return {
             'proxy-authorization': 'Basic '+
-                new Buffer(user+':'+password).toString('base64'),
+                Buffer.from(user+':'+password).toString('base64'),
         };
     }
     request(url, port, headers){
@@ -115,8 +115,7 @@ class Tracer {
             conf.proxy = `http://127.0.0.1:${port}`;
             if (this.proxies_running[port].opt.ssl)
                 conf.ca = ssl.ca.cert;
-            conf.headers = Object.assign({'x-hola-context':
-                'PROXY TESTER TOOL'}, headers);
+            conf.headers = headers;
         }
         const sp = etask(function*req_stats(){ yield this.wait(); });
         request(conf, (e, http_res)=>{
