@@ -1,19 +1,20 @@
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint node:true, esnext:true*/
-const _ = require('lodash');
 const lokijs = require('lokijs');
 const lfsa = require('lokijs/src/loki-fs-structured-adapter.js');
 const etask = require('../util/etask.js');
 const string = require('../util/string.js');
 const file = require('../util/file.js');
 const qw = string.qw;
+const log = require('./log.js');
 
 const E = module.exports = Loki;
 
-function Loki(path){
+function Loki(path, log_lvl){
     this.path = path;
     this.colls = {};
     this.db_names = qw`port status_code hostname protocol`;
+    this.log = log('loki', log_lvl);
 }
 
 E.prototype.prepare = etask._fn(function*loki_prepare(_this){
@@ -25,7 +26,7 @@ E.prototype.prepare = etask._fn(function*loki_prepare(_this){
         autoload: true,
         autoloadCallback: ()=>this.continue(),
         autosave: true,
-        autosaveInterval: 4000,
+        autosaveInterval: 60000,
     });
     yield this.wait();
     _this.db_names.forEach(db=>{
@@ -33,6 +34,18 @@ E.prototype.prepare = etask._fn(function*loki_prepare(_this){
             _this.loki.addCollection(db, {unique: [db], indices: ['reqs']});
         _this.colls[db] = _this.loki.getCollection(db);
     });
+});
+
+E.prototype.save = etask._fn(function*loki_save(_this){
+    const et = this;
+    _this.log.notice('Saving stats to the file');
+    _this.loki.saveDatabase(function(err){
+        if (!err)
+            return et.continue(true);
+        _this.log.error(err);
+        et.continue(err);
+    });
+    yield this.wait();
 });
 
 E.prototype.stats_clear = function(){
